requirements.txt
torch~=1.10.0
numpy~=1.21.4
pandas~=1.3.4
plotly~=5.1.0
scikit-learn~=1.0.1
scipy~=1.7.3
pytorch_lightning~=1.5.5
pyro-ppl~=1.8.0